+++ src/assets/js/protocol/protocol-lang-switcher.js
@@ -47,7 +47,7 @@ if (typeof Mozilla === 'undefined') {
      * @param {function} Custom callback for analytics.
      */
     LangSwitcher.init = function(callback) {
-        var language = document.querySelectorAll('.mzp-c-language-switcher-select');
+        var language = document.querySelectorAll('.mzp-js-language-switcher-select');
 
         for (var i = 0; i < language.length; i++) {
             language[i].setAttribute('data-previous-language', language[i].value);